Add a file to save commands of REPL on disk.
This feature allow to be able to store each command typed in the REPL.
It load them at each start of the REPL to be able to navigate through commands typed in an older REPL.
The commands are stored in the file specified in `base_cofing/config.toml`.
The REPL try to create the file if it doesn't exist each time.